requirements.txt
# Document Processor Pro - Python Dependencies
# Core functionality
PyPDF2==3.0.1
openpyxl==3.1.5

# Windows API integration
pywin32==310

# Custom icon support
Pillow>=9.0.0

# Build and distribution (optional)
pyinstaller>=5.0.0
